requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
python-dotenv==1.0.0
fake-useragent>=1.4.0
argparse==1.4.0
selenium==4.18.1
webdriver-manager==4.0.1 